refactor: 랜덤 이벤트 로직을 파일로 분리하기
