+++ src/basic/main.basic.js
@@ -1,7 +1,11 @@
 import { products, discountRateMap } from "./constants.js";
+import {
+  scheduleFlashSale,
+  scheduleRecommendationSale,
+} from "./helpers/scheduleTask.js";
 
 let productSelector, addToCartButton, cartItemList, cartTotal, stockStatus;
-let lastSel,
+let lastSelectedProductId,
   bonusPts = 0,
   finalTotal = 0,
   totalItemsInCart = 0;
@@ -55,49 +59,11 @@ function render() {
 }
 
 function triggerRandomSales() {
-  scheduleFlashSale();
-  scheduleRecommendationSale();
-}
-
-function scheduleFlashSale() {
-  const delay = Math.random() * 10000;
-
-  setTimeout(() => setInterval(flashSaleToRandomProduct, 30000), delay);
-}
-
-function scheduleRecommendationSale() {
-  const delay = Math.random() * 20000;
-
-  setTimeout(() => setInterval(suggestProductDiscount, 60000), delay);
-}
-
-function flashSaleToRandomProduct() {
-  const luckyItem = getRandomProduct();
-  if (Math.random() < 0.3 && luckyItem.units > 0) {
-    luckyItem.price = Math.round(luckyItem.price * 0.8);
-
-    alert(`번개세일! ${luckyItem.name}이(가) 20% 할인 중입니다!`);
-    updateProductSelector();
-  }
-}
-
-function suggestProductDiscount() {
-  if (!lastSel) return;
-
-  const suggestedItem = products.find(
-    (item) => item.id !== lastSel && item.units > 0,
-  );
-  if (suggestedItem) {
-    suggestedItem.price = Math.round(suggestedItem.price * 0.95);
-    alert(
-      `${suggestedItem.name}은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!`,
-    );
-    updateProductSelector();
-  }
-}
-
-function getRandomProduct() {
-  return products[Math.floor(Math.random() * products.length)];
+  scheduleFlashSale({ onSale: updateProductSelector });
+  scheduleRecommendationSale({
+    productId: lastSelectedProductId,
+    onSale: updateProductSelector,
+  });
 }
 
 function addEventListener() {
@@ -113,7 +79,7 @@ function addEventListener() {
         createNewCartItem(selectedProduct);
       }
       calculateCart();
-      lastSel = selectedProduct.id;
+      lastSelectedProductId = selectedProduct.id;
     }
   });
